usdt: enable tests on arm64
Many of the amd64 unit tests also apply to arm64 now that we have arm64 support
for USDT. The only test that was modified was TestUsdtArgs. The usdt_sib check
was made conditional on the runtime being amd64. Otherwise, everything else in
this commit is ether a file rename or an exact copy/paste.
